Refactored edit modal to include both add and edit functionality:
* Moved functionality of adding user to the new AddEditModal
* Removed AddModal
