--- web/src/components/admin-dashboard-overview/manage-users/add-modal/add-modal.types.ts
@@ -1,15 +1,0 @@
-import { IUserData } from "../../../../components/admin-dashboard-overview/admin-dashboard-overview.types";
-import { KeyedMutator } from "swr";
-
-export interface AddModalProps {
-    open: boolean;
-    onClose: () => void;
-    mutateUsers: KeyedMutator<IUserData[]>;
-}
-
-export interface IAddUserForm {
-    firstName: string;
-    lastName: string;
-    designation: string;
-    email: string;
-}
