--- web/src/components/admin-dashboard-overview/manage-users/add-modal/add-modal.utils.ts
@@ -1,8 +1,0 @@
-export const getDefaultValues = () => {
-    return {
-        firstName: "",
-        lastName: "",
-        designation: "",
-        email: "",
-    };
-};
